# src/InactivePDF.Api/wwwroot/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InactivePDF API",
    "version": "0.1.0-beta",
    "description": "Self-hosted document-to-PDF conversion API. Use an administrator bearer token for administration routes or a scoped integration key for conversion and owned-job routes."
  },
  "servers": [{ "url": "/", "description": "Current InactivePDF instance" }],
...
  "paths": {
...
    "/v1/convert-file": { "post": { "tags": ["Conversion"], "summary": "Convert one file synchronously", "security": [{ "integrationBearer": [] }, { "administratorBearer": [] }], "requestBody": { "$ref": "#/components/requestBodies/ConversionMultipart" }, "responses": { "200": { "$ref": "#/components/responses/Pdf" }, "default": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/create-text-pdf": { "post": { "tags": ["Conversion"], "summary": "Create a PDF from text", "security": [{ "integrationBearer": [] }, { "administratorBearer": [] }], "parameters": [{ "name": "watermarkProfile", "in": "query", "schema": { "type": "string" }, "description": "Optional saved watermark profile; direct watermark options in the body take precedence." }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TextPdfRequest" }, "example": { "text": "Hello from InactivePDF", "profile": "archive" } } } }, "responses": { "200": { "$ref": "#/components/responses/Pdf" }, "default": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/convert-and-merge": { "post": { "tags": ["Conversion"], "summary": "Convert and merge files synchronously", "security": [{ "integrationBearer": [] }, { "administratorBearer": [] }], "requestBody": { "$ref": "#/components/requestBodies/ConversionMultipart" }, "responses": { "200": { "$ref": "#/components/responses/Pdf" }, "default": { "$ref": "#/components/responses/Problem" } } } },
...
  },
  "components": {
    "securitySchemes": {
      "administratorBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "administrator token" },
      "integrationBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "integration API key" }
    },
...
    "requestBodies": {
      "ConversionMultipart": { "required": true, "content": { "multipart/form-data": { "schema": { "$ref": "#/components/schemas/ConversionForm" } } } },
...
    },
    "schemas": {
...
      "ConversionForm": { "type": "object", "required": ["file"], "properties": { "mode": { "type": "string", "enum": ["sync", "async"], "default": "async" }, "profile": { "type": "string", "default": "archive" }, "watermarkProfile": { "type": "string" }, "watermark": { "type": "string", "description": "JSON-encoded WatermarkOptions for multipart requests. Direct options take precedence over watermarkProfile." }, "file": { "type": "array", "items": { "type": "string", "format": "binary" } } } },
...
      "TextPdfRequest": { "type": "object", "required": ["text"], "properties": { "text": { "type": "string" }, "profile": { "type": "string", "default": "archive" }, "watermarkProfile": { "type": "string" }, "watermark": { "type": "object" } } },
...
      "Problem": { "type": "object", "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "requestId": { "type": "string" }, "retryable": { "type": "boolean" } } }
    },
    "responses": {
      "Pdf": { "description": "Generated PDF", "content": { "application/pdf": { "schema": { "type": "string", "format": "binary" } } } },
...
      "Problem": { "description": "Structured error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
...
    }
  }
}